requirements.txt
boto3==1.21.18
botocore==1.24.18
jmespath==0.10.0
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.8
